to u32

+++ src/homeworks.rs
@@ -16,9 +16,9 @@ pub struct Comment {
     /// Equal to `creation_ts` if the comment has never been edited.
     pub last_edited_ts: i64,
     /// Number of upvotes 
-    pub upvotes: i64,
+    pub upvotes: u32,
     /// Number of downvotes.
-    pub downvotes: u64,
+    pub downvotes: u32,
     /// The vote of the current user.
     /// -1, 0 or 1.
     pub vote: i8,
